refactor: rename Shape namespace with Renderer
