+++ Classes/Renderer/RectShapeRenderer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Blueways\BwFocuspointImages\Shape;
+namespace Blueways\BwFocuspointImages\Renderer;
 
 final readonly class RectShapeRenderer implements ShapeRendererInterface
 {
